cleanup(core): extract the migration execution engine into a separate module
The per-migration execution engine (runNxOrAngularMigration and its
callees, ChangedDepInstaller, runInstall and its error classification,
collection and implementation resolution) lived inside the 4100-line
migrate.ts alongside the planning phase and the multi-migration run
loop. Move it verbatim into execute-migration.ts so upcoming consumers
can use it without pulling in the rest of migrate.ts.

migrate.ts re-exports the whole engine surface, so existing importers
(migrate-ui-api.ts, run-migration-process.js) are unaffected. The new
migrate-contracts.spec.ts pins that re-export set and the contracts
those consumers rely on.
